Rename `dsl_spec` into `dsl_spec_helper` so it doesn't get executed automnatically
